Split authentication mutations
