+++ saleor/graphql/account/tests/mutations/test_tokens_deactivate_all.py
@@ -5,7 +5,7 @@ from freezegun import freeze_time
 from .....account.error_codes import AccountErrorCode
 from .....core.jwt import create_access_token, create_refresh_token, jwt_decode
 from ....tests.utils import get_graphql_content
-from ...mutations.authentication import _get_new_csrf_token
+from ...mutations.authentication.utils import _get_new_csrf_token
 from .test_token_refresh import MUTATION_TOKEN_REFRESH
 
 MUTATION_DEACTIVATE_ALL_USER_TOKENS = """
